+++ sources/src/main/java/com/google/solutions/jitaccess/web/rest/PolicyResource.java
@@ -25,7 +25,7 @@ import com.google.common.base.Preconditions;
 import com.google.solutions.jitaccess.catalog.policy.IamRoleBinding;
 import com.google.solutions.jitaccess.catalog.policy.Policy;
 import com.google.solutions.jitaccess.catalog.policy.PolicyDocument;
-import com.google.solutions.jitaccess.apis.IamRoleResolver;
+import com.google.solutions.jitaccess.auth.IamRoleResolver;
 import com.google.solutions.jitaccess.catalog.policy.PolicyIssue;
 import com.google.solutions.jitaccess.util.Cast;
 import com.google.solutions.jitaccess.util.MoreStrings;
